[Codegen] Push up memref reshapes before empty tensor elimination
