+++ compiler/src/iree/compiler/Codegen/Common/BlockDynamicDimensions.cpp
@@ -176,20 +176,6 @@ blockDynamicDimensionsOfValue(RewriterBase &rewriter,
   return ReshapeOps{expandShapeOp, collapseShapeOp};
 }
 
-void moveUpMemrefReshapeOps(RewriterBase &rewriter, Operation *op) {
-  DominanceInfo domInfo(op);
-  SmallVector<Operation *> reshapeOps;
-  op->walk([&](Operation *nestedOp) {
-    if (isa<memref::ExpandShapeOp, memref::CollapseShapeOp>(nestedOp)) {
-      reshapeOps.push_back(nestedOp);
-    }
-  });
-
-  for (Operation *reshapeOp : reshapeOps) {
-    moveOpAfterLastOperand(rewriter, domInfo, reshapeOp);
-  }
-}
-
 //===---------------------------------------------------------------------===//
 // Methods for blocking operands of operations
 //===---------------------------------------------------------------------===//
